Split Agda.Utils.Range from Agda.Interaction.Highlighting.Range
Separate out the generic, non-Agda specific stuff that is used by `Agda.Utils.RangeMap`.
