+++ src/full/Agda/Interaction/Highlighting/Precise.hs
@@ -45,7 +45,7 @@ import qualified Agda.Syntax.Common   as Common
 import Agda.Syntax.TopLevelModuleName
 import Agda.Syntax.Scope.Base                   ( KindOfName(..) )
 
-import Agda.Interaction.Highlighting.Range
+import Agda.Utils.Range
 
 import qualified Agda.Utils.List1 as List1
 import Agda.Utils.Maybe
